move p2p stats related test cases into single module
