+++ jormungandr-scenario-tests/src/test/features/p2p/stats.rs
